Minor:Refactoring checklist component.
