+++ package.json
@@ -22,9 +22,9 @@
     "@angular/platform-browser-dynamic": "^4.2.4",
     "@angular/router": "^4.2.4",
     "angular-animate": "^1.6.6",
-    "chart.js": "^2.5.0",
     "angular-bootstrap-md": "^4.3.5",
     "bootstrap": "^3.3.7",
+    "chart.js": "^2.5.0",
     "core-js": "^2.4.1",
     "font-awesome": "^4.7.0",
     "hammerjs": "^2.0.8",
