+++ src/app/app.module.ts
@@ -19,7 +19,7 @@ import { ChecklistthreeComponent } from './checklists/checklistthree/checklistth
 import { SchedulingHomeComponent } from './scheduling/scheduling-home/scheduling-home.component';
 import { SchedulingComponent } from './scheduling/scheduling-meeting/scheduling-meeting.component';
 import { SchedulingModule } from './scheduling/scheduling-meeting/scheduling-meeting.module';
-import { ChecklistModule } from './checklists/checklist.module';
+import { ChecklistModule } from './checklist/checklist.module';
 import { SchedulingVisitComponent } from './scheduling/scheduling-visit/scheduling-visit.component';
 import { ROUTES } from './app.routes';
 import { SigninComponent } from './layouts/home/signin/signin.component';
