+++ src/app/app.routes.ts
@@ -9,7 +9,7 @@ import { SigninComponent } from './layouts/home/signin/signin.component';
 import {ChecklistoneComponent} from './checklists/checklistone/checklistone.component';
 import {ChecklisttwoComponent} from './checklists/checklisttwo/checklisttwo.component';
 import {ChecklistthreeComponent} from './checklists/checklistthree/checklistthree.component'
-import {ChecklistComponent} from './checklists/checklist.component'
+import {ChecklistComponent} from './checklist/checklist.component'
 export const ROUTES: Routes = [
   { path: '', redirectTo: '/home', pathMatch: 'full' },
   { path: 'home', component: HomeComponent },
