--- src/app/checklists/checklistsModels/checklist.model.ts
@@ -1,12 +1,0 @@
-import { SchoolModel } from './school.model';
-//This class represents the question form. It includes the three sectors.
-export class ChecklistModel{
-
-    private idChecklist: number;
-    private date: Date;
-    private responsibleInformation: string;
-    private cookSchool: string;
-    private school: SchoolModel;
-    
-    constructor(){}
-}
