+++ src/app/models/checklist-kitchen.model.ts
@@ -1,0 +1,9 @@
+import { Form } from './form.interf';
+
+export class ChecklistKitchen{
+
+    constructor(private nameCook: string,
+                private form : Array<Form>,
+                private foodRatio : string    
+    ){}
+}
