+++ src/app/models/form.interf.ts
@@ -1,0 +1,5 @@
+export interface Form{
+    
+    question: string;
+    answer: boolean;
+}
