+++ src/app/models/responsible-information.interf.ts
@@ -1,0 +1,5 @@
+export interface ResponsibleInformation{
+    
+    name : string;
+    function : string;
+}
