+++ src/app/models/school.model.ts
